improved reply fetching, along with fetching reply updates
