+++ src/routes/timeline/index.js
@@ -14,16 +14,15 @@ router.get('/new', require('./new')) // get new posts since last fetched timelin
  */
 router.get('/', async (req, res, next) => {
   // Validate firstPostId
-  if (req.query.firstPostId && !mongoose.Types.ObjectId.isValid(req.query.firstPostId)) res.sendStatus(404) // Unknown first post id
+  if (req.query.firstPostId && !mongoose.Types.ObjectId.isValid(req.query.firstPostId)) return res.sendStatus(404) // Unknown first post id
 
   // Normalize options
   let options = {
-    amount: req.query.amount ? parseInt(req.query.amount) : 5, // default 25
+    amount: req.query.amount ? parseInt(req.query.amount) : 25, // default 25
     offset: req.query.offset ? parseInt(req.query.offset) : 0 // default 0
   }
   // Validate options
-  if (options.amount <= 0 || options.amount > 25) return res.sendStatus(401) // Unauthorized
-  if (options.offset < 0) return res.sendStatus(404) // Page not found 404
+  if (options.amount <= 0 || options.amount > 25 || options.offset < 0) return res.sendStatus(422) // Unauthorized
 
   // Get who the user is following
   let following
